Transferred Everything to Main
I was running into too many issues with the GUI I created, so I chose to run the code through the console. I would rather have the calculator working properly in the console than have it not work at all in the GUI. So I am no longer using a GUI and all the code is now in my Main.java file.